Show parameters derived from LCSC descrtiption
This adds an "LCSC Params" field to the footprint and part selection lists
that makes it much easier to double-check the assignments or select the
right part. Particularly for passives, LCSC buries the important data (like
the resistance of a resistor) in the middle of the description field. This
pulls those values out where you can see them.